requirements.txt
uvicorn==0.42.0
pydantic==2.12.5
paho-mqtt==2.1.0
psycopg==3.3.3
psycopg-binary==3.3.3
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.1.3
fastapi==0.135.1
sqlalchemy==2.0.48
python-dotenv==1.2.2